Store GKE credentials in buildfolder/.config/gcloud

+++ backend/gke/kubeconfig.sh
@@ -8,6 +8,14 @@ if [ ! -f "$KUBECFG" ]; then
     exit 1
 fi
 
+# check gcloud credentials:
+info "Using creds from GKE_CRED_JSON…"
+gcloud auth revoke 2>/dev/null || true
+gcloud auth activate-service-account --key-file "$GKE_CRED_JSON"
+if [[ $(gcloud auth list  --format="value(account)" | wc -l ) -le 0 ]]; then
+    err "GKE_CRED_JSON creds don't authenticate, aborting" && exit 1
+fi
+
 cp "$KUBECFG" kubeconfig
 
 # kubeconfig gets hardcoded paths for gcloud bin, reevaluate them:
